Created seperate classes for responses, request, created seperate services, and a EncodingServiceFactory.cs.
This simplifies encoding requests to such level, that a simple /api/Encoding call can distinguish what type of message is trying to be encoded, and then allocates a dedicated service to process that request. No need for differing httppost requests, simplifies the controller to only be the link between front-end and back-end. all the calculations are made in the dedicated service.